fix abci query
